requirements.txt
# Dependencias de Python
requests==2.31.0
pandas==2.2.1
numpy==1.26.4
fastparquet==2024.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1


